Resolves GH-1: Fix formatting of header to match GitHub documentation
Previously, the security header generated from a provided webhook secret
only consisted of the generated value. GitHub documents that when a
webhook is sent, they send `sha1=`, followed by the value. This
mis-match was causing mocked webhook requests using secrets to fail.

This change adds the required prefix when a secret is provided as part
of the generated value